move bitset to its own file
